+++ Libplanet/Crypto/DefaultCryptoBackend.cs
@@ -7,9 +7,10 @@ using Org.BouncyCastle.Math;
 
 namespace Libplanet.Crypto
 {
-    public class DefaultCryptoBackend : ICryptoBackend
+    public class DefaultCryptoBackend<T> : ICryptoBackend<T>
+        where T : HashAlgorithm
     {
-        public byte[] Sign(HashDigest<SHA256> messageHash, PrivateKey privateKey)
+        public byte[] Sign(HashDigest<T> messageHash, PrivateKey privateKey)
         {
             var h = new Sha256Digest();
             var kCalculator = new HMacDsaKCalculator(h);
@@ -34,7 +35,7 @@ namespace Libplanet.Crypto
         }
 
         public bool Verify(
-            HashDigest<SHA256> messageHash,
+            HashDigest<T> messageHash,
             byte[] signature,
             PublicKey publicKey)
         {